fix bug

+++ src/mouvement/util_mouv.c
@@ -6,7 +6,7 @@
 /*   By: iouazzan <iouazzan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/23 13:49:40 by iouazzan          #+#    #+#             */
-/*   Updated: 2022/12/02 17:47:48 by iouazzan         ###   ########.fr       */
+/*   Updated: 2022/12/02 21:51:11 by iouazzan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int	check_next_step(t_data *data, int x, int y)
 	y /= 64;
 	if (data->gm->map[x][y] != '1')
 	{
-		if (x != (data->pp_x / 64) && y != (data->pp_y / 64))
+		if (x != (((int)data->pp_x) / 64) && y != ((int)data->pp_y / 64))
 			return (check_sides(data, x, y));
 		return (0);
 	}
